Add a test in macaw-refinement for the corresponding macaw-symbolic code
The code in macaw-symbolic looks in the external resolutions field to generate
Crucible code performing indirect call dispatch based on computed IP values.
This test case makes sure that we actually
1. Resolve the control flow correctly
2. Correctly use it to build a Crucible CFG

The test is kind of rudimentary in that it doesn't try to prove anything in
particular - it just makes sure that the function has no unresolved control flow
that causes an abort.