Don't generate semantics for ARM vector instructions
Instead, just represent them as uninterpreted.  The full set (including the
vector instructions) was way too large to compile.  This is a compromise until
we generate smaller code.

To accomplish this, there are now arch-specific filters to inhibit codegen for
library functions and opcodes.

To plumb these through, we now have a Macaw SemMC TH config object to bundle up
all of the config parameters.  This reduces the number of parameters to a number
of functions, hopefully improving readability.

NOTE: There are still some cases to write for traversable/show instances for the
new ARM function constructors.