Fix handling thumb state (PSTATE_T) between blocks and at function boundaries
This change uses the block precondition support to propagate the value of
`PSTATE_T` between blocks in a function. It also sets the `PSTATE_T` value for a
function based on the low bit in the function address.

Note that we are currently forced to assume that `PSTATE_T` is preserved across
calls, even though that isn't strictly necessary.