Add a testing framework for macaw-symbolic
This module provides a few functions to make it easier to write test suites for
the architecture-specific macaw-symbolic backends.  These functions try to not
be too prescriptive - they provide some tools to load binaries and set up the
symbolic execution engine, as well as run it in a way suitable for testing.