Fix a bug in the AArch32-specific simplifier (#188)
Some important simplifications for classification were failing to fire because
other simplifications fired first, short circuiting the search.  It turns out
that more than one rule may apply at any given step (and it is important to
apply all of the rules that can be applied).  This commit modifies the
simplifier to apply rules until saturation.