Remove the TOC code from the ppc32 interface
The most common ppc32 ABIs do not use a Table of Contents (TOC). It is possible
to have them do so with the right compilation flags, but it is not something we
should be expressing at the type level.

For now, we are totally removing the TOC handling for ppc32. If we need to
support those uncommon ABIs, we will do so at the data level.