Revise system call handling
This new translation requires some tight coupling between macaw-x86 and
macaw-x86-symbolic in order to arrange for the proper data flows. It uses a
simple arch-specific function to represent system calls in macaw-x86, but
translates it into two other primitives in macaw-x86-symbolic that faithfully
implement system call semantics:

1. Look up the function handle for the syscall
2. Invoke the function (in a way that flows the updates registers back to the
calling context)

See the note in Data.Macaw.X86.Symbolic for details.

This commit removes some of the other attempts at adding syscall support. It
adds shared infrastructure for looking up system calls in macaw-symbolic, but
delegates the use of that primitive to the architecture-specific backends.