Add support for FSBASE and GSBASE on x86_64
This change adds support for the FSBASE and GSBASE registers on x86_64.
It also lays the groundwork for supporting additional special segment
registers on other architectures.  It does this by adding a callback
`LookupSegmentBasePointer` that takes an `ArchFn` indicating which
segment register is being read, and returns a base pointer to that
memory segment.