Enable ArchTermStmts to contain values
Before, `ArchTermStmt` was only parameterized by a state thread (`ids`).  This
meant that they could not contain macaw (or crucible) values.  Some work on
AArch32 requires being able to store condition values in arch terminators (to
support conditional returns). This change modifies the `ArchTermStmt` to enable
this, which requires a bit of plumbing through various definitions and some
extra instances.

In support of actually using this, it also became necessary to plumb fallthrough
block labels through the architecture-specific terminator translation in
macaw-symbolic.

Note that this change was overdue, as the PowerPC backend was storing macaw
values in a way that would have rendered them unusable in the macaw-ppc-symbolic
translation, had any interpretation been provided.  These new changes will
enable a handler to be written for the conditional PowerPC trap instructions.

PowerPC, x86, and ARM have been updated.