arm: Add transfer functions for division
Add support for abstract interpretation of division operations in the ARM
backend (when the operands to division are concretely known). Note that this
only covers SDIV and UDIV; I couldn't find documentation on what the exact
semantics of the *REM operations are.

This commit also adds extended documentation on the semantics of these
operations.