Fix tail call classification
The tail call classifier came after the jump classifier, which was a problem
because it is less strict than the tail call classifier, meaning it would always
fire.  This commit moves direct jump to be the last classifier applied, giving
the others a chance.

Includes a test case in the ARM backend.

Fixes #285