Fix some test expected values
They changed because a few blocks are now classified as tail calls that were
plain jumps before.  They really could be considered either.  I think it would
be nice if these could be classified as jumps instead, but the reason they are
flagged as tail calls is mostly down to the fact that their surrounding context
is so simple that either interpretation works.

Correcting this would require some heuristics based on additional analysis passes.