add additional case to PPC matchReturn
The PPC return sequence (from the PPC semantics)
implicitly wraps the address in shift left/shift right to mask
the lower two bits, e.g.:

      r15 := (bv_shr r13 (0x2 :: [32]))
      r16 := (trunc r15 30)
      r17 := (uext r16 32)
      r18 := (bv_shl r17 (0x2 :: [32]))

Prior to this patch, this was the only pattern that was
handled by matchReturn, and thus considered a valid return
address.

We observed in some cases that binaries compiled with -O2
would have the following expression appear in the link
register prior to a tail function call:

      r8 := (bv_add r1_0 (0x14 :: [32]))
      r9 := read_mem r8 (bvbe4)

Where r9 is read directly off the stack without any
additional shifting.

This second pattern causes the tail call classifier to fail,
because matchReturn did not recognize this sequence.

This change allows matchReturn to succeed on the latter case,
and therefore causes the tail call classifier to succeed as
expected.