Data.Macaw.PPC.Identify: adds a strictness flag to matchRead to allow more permissive matching of returns in matchReturn
This flag gives us a workaround for cases when Macaw has lost abstract
domain information and can't determine that the IP is a return address.
Specifically, if the value has the expected shift/truncate/extend
structure, then we assume that it is a return.

There's a risk in this approach because Macaw might now incorrectly
identify some blocks as terminating with a return.