ArchitectureInfo: simplify checkForReturnAddr return type
This change updates the checkForReturnAddr record field type so that
rather than returning a value that was never used (and was always
going to be ReturnAddr anyway), the function now returns unit to
indicate success or fails via "fail" otherwise. This resulted in a
simplification of the various spots where checkForReturnAddr was called
or was implemented for each architecture.

As part of this change, tryClassifier was removed since that was only
used as part of the previous implementation.