symbolic: Remove redundant unwrapping/rewrapping of register struct
There was a lot of redundant unwrapping and re-wrapping of the
assignment of `TypeRepr`s to the architecture register context into and
out of a `StructRepr`. Remove this.

Also, add an hlint configuration file to help avoid this in the future,
and a Github Actions workflow to enforce this in CI.