macaw-loader: Make entryPoints return a possibly empty list
It is not guaranteed that an ELF file's `e_entry` value will always correspond
to a valid entry point address, which means that we cannot guarantee that the
list of entry points that `macaw-loader`'s `entryPoints` computes will always
be non-empty. This patch changes the type of `entryPoints` to return `[]`
instead of `NonEmpty` and makes it so that failing to resolve the `e_entry`
address in `entryPoints` does not throw an exception.

Fixes #482.