symbolic: Fix soundness bug in lazy memory model involving path conditions
Fixes #491. The problem was that assumptions about the contents of
memory were added on a per-path basis, when they should have been
"global" or "top-level" assumptions, available for proving *all* goals.

Here's a my understanding of the end-to-end flow of the bug.

At the beginning of symbolic execution, we have an empty `GoalCollector`
in the Crucible `Backend`: `TopCollector (Seq.fromList [])`. Then, we
make some assertions which eventually hit this case:

https://github.com/GaloisInc/crucible/blob/52ea65bcf89144d7e1e74a00c989a7040fc7924d/crucible/src/Lang/Crucible/Backend/ProofGoals.hs#L209

resulting in `TopCollector (Seq.fromList [Prove g0, ...])`. Then
we hit a symbolic branch, resulting in pushing a frame and perhaps
some more goals and assumptions, e.g., perhaps `CollectingGoals gs
(CollectingAssumptions as (CollectorFrame f (TopCollector {- same as
before -})))`.

Now, we read from a global and generate an assumption that some piece
of global memory points to its initializer. We need to make this new
assumption available when dispatching *all* of the proof obligations
that we have generated so far. How can we do this?

We call `gcAddTopLevelAssumption`, which results in
```
CollectingGoals gs (CollectingAssumptions as (CollectorFrame f (CollectingAssumptions newGlobalMemAssumption (TopCollector (Seq.fromList [Assuming newGlobalMemAssumption (Prove g0), ...])))))
```
(For the "duplication" of `newGlobalMemAssumption`, see the comment on
`gcAddTopLevelAssumption`.)

So why didn't `gcAddTopLevelAssumption` exist before? Because we never
had a use-case for it. During normal symbolic execution, you generally
only add new assumptions for the path you're on (i.e., in the frames
you're in). There is generally no need to "go back" and add assumptions
that are available for all goals. If you have such assumptions, you
usually just add them "up front", which IIUC is what the non-lazy Macaw
memory model does.